Discard majority of pointer usage through codebase
I was native to use pointers a lot of times throughout the codebase.
None of the arguments to functions are "large" (e.g. structs are not what
I consider to be large), and most do not "need" to modify there
arguments.

The only case where this is not true is ramunderprice, with one of
its methods needing to modify the receiver struct. Hence, to be
consistent, all methods of the rule type will be pointer receivers.