Integrate smtp for matches, change postThreshold
To clarify, I am talking about the atPostThreshold method. I noticed
sometimes the number of posts held by the queue would be greater than
what the postThreshold was expecting.

matchPosts was changed from a slice to a mapping of rule name to post
so that way in email I know the rule that matched for a post.